Refactor dir/file structure.
