+++ src/js/features/Servers/FavoriteToggle.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { connect } from "react-redux";
-import { toggleFavoriteServer } from "../UserInterface/slice.js";
+import { toggleFavoriteServer } from "./../../uiSlice.js";
 
 const mapStateToProps = (state, ownProps) => ({
   isEnabled: state.ui.favorites.servers.includes(ownProps.serverAddress),
