+++ src/js/store.js
@@ -1,13 +1,13 @@
 import { configureStore } from "@reduxjs/toolkit";
 //import { setupListeners } from '@reduxjs/toolkit/query'
 import { qwsSlice } from "./services/qws";
-import uiReducer from "./features/UserInterface/slice.js";
+import uiSlice from "./uiSlice.js";
 
 export const store = configureStore({
   reducer: {
     // Add the generated reducer as a specific top-level slice
     [qwsSlice.reducerPath]: qwsSlice.reducer,
-    ui: uiReducer,
+    ui: uiSlice.reducer,
   },
   // Adding the api middleware enables caching, invalidation, polling,
   // and other useful features of `rtk-query`.
