+++ src/js/util/clipboard.js
