+++ src/js/util/storage.js
