+++ src/js/util/text.js
